Added practice in testing

+++ Spring/spring-testing/src/main/java/com/example/demo/student/StudentController.java
@@ -24,7 +24,7 @@ public class StudentController {
     }
 
     @DeleteMapping(path = "{studentId}")
-    public void deleteStudent(@PathVariable("studentId") Long studentId) {
+    public void deleteStudent(@PathVariable("studentId") int studentId) {
         studentService.deleteStudent(studentId);
     }
 }
